requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
onnx>=1.14.0
onnxruntime>=1.16.0
matplotlib>=3.7.0
Pillow>=10.0.0
tqdm>=4.65.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
sam-2 @ git+https://github.com/facebookresearch/sam2.git
lightning-sam @ git+https://github.com/luca-medeiros/lightning-sam.git